requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-SQLAlchemy==3.0.5
WTForms==3.0.1
requests==2.31.0
python-dotenv==1.0.1 
Werkzeug==2.3.7
email_validator==2.1.1
google-generativeai==0.8.5
Jinja2==3.1.6
SQLAlchemy==2.0.41
mistune==2.0.5